+++ Maddori.Apple/Maddori.Apple/Global/Extension/UIViewController+Extension.swift
@@ -68,15 +68,3 @@ extension UIViewController {
         view.endEditing(true)
     }
 }
-
-extension UIView {
-    func setGradient(colorTop: UIColor, colorBottom: UIColor) {
-        let gradient: CAGradientLayer = CAGradientLayer()
-        gradient.colors = [colorTop.cgColor, colorBottom.cgColor]
-        gradient.locations = [0.0, 1,0]
-        gradient.startPoint = CGPoint(x: 0.0, y: 0.0)
-        gradient.endPoint = CGPoint(x: 0.0, y: 1.0)
-        gradient.frame = bounds
-        layer.addSublayer(gradient)
-    }
-}
